backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
openai
pytest
pytest-asyncio
pytest-cov
httpx
requests
coverage